% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arp.R
\name{arp}
\alias{arp}
\title{ARP: Accuracy, Recall, Precision}
\usage{
arp(y_hat, y_actual)
}
\arguments{
\item{y_hat}{Predicted values}

\item{y_actual}{Actual/observed values}
}
\description{
Computes accuracy metrics of a binary classification model: Accuracy, Recall, Precision. Returns a matrix. Kindly provided by Agata Dziwok.
}
\examples{
arp()
}
\keyword{Accuracy,}
\keyword{Classification,}
\keyword{Precision}
\keyword{Recall,}
