% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta.R
\name{Beta}
\alias{Beta}
\title{Beta}
\usage{
Beta(ret, t_ret, rf_ret)
}
\arguments{
\item{ret}{A return of an asset}

\item{t_ret}{Return of the tangency portfolio}

\item{rf_ret}{Return of risk free asset}
}
\description{
Computes beta, given assets's return, return of tangency portfolio and riskless return
}
\examples{
Beta()
}
\keyword{Beta,}
\keyword{Finance,}
\keyword{Portfolio}
\keyword{Tangency}
