% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(cov_m)
}
\arguments{
\item{cov_m}{Covariance matrix}
}
\description{
Returns correlation matrix given covariance matrix
}
\examples{
correlation()
}
\keyword{Correlation,}
\keyword{Covariance,}
\keyword{Statistics,}
\keyword{slovakgirl}
