% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolioReturn.R
\name{portfolioReturn}
\alias{portfolioReturn}
\title{Portfolio Return}
\usage{
portfolioReturn(w_vec, return_vec)
}
\arguments{
\item{w_vec}{Vector of assets' weights}

\item{return_vec}{Vector of assets' returns}
}
\description{
Outputs a return of a portfolio given a vector of its assets' returns and a vector of their respective weights
}
\examples{
portfolioReturn()
}
\keyword{Finance}
\keyword{Return,}
