% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portVar.R
\name{portVar}
\alias{portVar}
\title{Portfolio Variance}
\usage{
portVar(cov_m, w_vec)
}
\arguments{
\item{cov_m}{Covariance matrix of portfolio's assets' returns}

\item{w_vec}{Vector of assets' weights}
}
\description{
Calculates variance of a portfolio, given its assets' covariance matrix and weights
}
\examples{
portVar()
}
\keyword{Finance,}
\keyword{Portfolio}
\keyword{Variance,}
