% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeReturns.R
\name{relativeReturns}
\alias{relativeReturns}
\title{Relative returns}
\usage{
relativeReturns(vec)
}
\arguments{
\item{vec}{Vector of asset prices}
}
\description{
Returns a vector of relative returns for a given vector of asset prices
}
\examples{
relativeReturns()
}
\keyword{Finance,}
\keyword{Relative}
\keyword{Return,}
\keyword{return}
