% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(y_hat, y_actual)
}
\arguments{
\item{y_hat}{Vector of predicted values}

\item{y_actual}{Vector of observed/actual values}
}
\description{
Returns root mean squared error for given vectors of observed and predicted values
}
\examples{
rmse()
}
\keyword{RMSE,}
\keyword{Root}
\keyword{error,}
\keyword{mean}
\keyword{regression,}
\keyword{squared}
