% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangencyWeights.R
\name{tangencyWeights}
\alias{tangencyWeights}
\title{Tangency weights}
\usage{
tangencyWeights(cov_m, returns_v, rf_rate)
}
\arguments{
\item{cov_m}{Covariance matrix of assets' returns}

\item{returns_v}{Vector of assets' returns}

\item{rf_rate}{Risk free rate}
}
\description{
Returns a vector of assets' weights in a tangency portfolio
}
\examples{
tangencyWeights()
}
\keyword{Finance,}
\keyword{Tangency}
\keyword{portfolio}
